# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PropConnect AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/properties": {
      "get": {
        "tags": [
          "properties"
        ],
        "summary": "List Properties",
        "operationId": "list_properties_api_properties_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search address, city or ZIP",
              "title": "Q"
            },
            "description": "Search address, city or ZIP"
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Price"
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "name": "min_beds",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Beds"
            }
          },
          {
            "name": "min_baths",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Baths"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PropertyStatus"
                },
                {
                  "type": "null"
                }
              ],
              "default": "for_sale",
              "title": "Status"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "price_asc",
                "price_desc",
                "newest",
                "size_desc"
              ],
              "type": "string",
              "default": "price_asc",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "properties"
        ],
        "summary": "Create Property",
        "operationId": "create_property_api_properties_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/properties/cities": {
      "get": {
        "tags": [
          "properties"
        ],
        "summary": "List Cities",
        "operationId": "list_cities_api_properties_cities_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 60,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityCount"
                  },
                  "title": "Response List Cities Api Properties Cities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/properties/{property_id}": {
      "get": {
        "tags": [
          "properties"
        ],
        "summary": "Read Property",
        "operationId": "read_property_api_properties__property_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "properties"
        ],
        "summary": "Update Property",
        "operationId": "update_property_api_properties__property_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Property Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "properties"
        ],
        "summary": "Delete Property",
        "operationId": "delete_property_api_properties__property_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CityCount": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "city",
          "count"
        ],
        "title": "CityCount"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PropertyCreate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PropertyStatus",
            "default": "for_sale"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "bedrooms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms"
          },
          "bathrooms": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathrooms"
          },
          "house_size_sqft": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "House Size Sqft"
          },
          "acre_lot": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Acre Lot"
          },
          "street_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Street Address"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "City"
          },
          "state": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "State"
          },
          "zip_code": {
            "type": "string",
            "pattern": "^\\d{5}$",
            "title": "Zip Code"
          },
          "property_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Type"
          }
        },
        "type": "object",
        "required": [
          "city",
          "state",
          "zip_code"
        ],
        "title": "PropertyCreate"
      },
...
      "PropertyPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PropertyRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PropertyPage"
      },
      "PropertyRead": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PropertyStatus",
            "default": "for_sale"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "bedrooms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms"
          },
          "bathrooms": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathrooms"
          },
          "house_size_sqft": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "House Size Sqft"
          },
          "acre_lot": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Acre Lot"
          },
          "street_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Street Address"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "City"
          },
          "state": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "State"
          },
          "zip_code": {
            "type": "string",
            "pattern": "^\\d{5}$",
            "title": "Zip Code"
          },
          "property_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Type"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "prev_sold_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Sold Date"
          }
        },
        "type": "object",
        "required": [
          "city",
          "state",
          "zip_code",
          "id"
        ],
        "title": "PropertyRead"
      },
      "PropertyStatus": {
        "type": "string",
        "enum": [
          "for_sale",
          "sold",
          "ready_to_build",
          "off_market"
        ],
        "title": "PropertyStatus"
      },
      "PropertyUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "bedrooms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms"
          },
          "bathrooms": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathrooms"
          },
          "house_size_sqft": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "House Size Sqft"
          },
          "street_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Street Address"
          },
          "property_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Type"
          }
        },
        "type": "object",
        "title": "PropertyUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}